[stage1] liveness-aware reduction-chunk decision + welford huge-M cap & fp32 stats
Stage 1 of the reduction-seed heuristic stack.

Part A — fold the reduction-chunk size AND the persistent verdict into one budgeted
`_reduction_rblock` returning (r_block, persistent); behavior-preserving (zero config diff).

Part B — a walker liveness fact `ReductionFact.body_live_tiles` (peak simultaneously-live
rdim-shaped tiles, computed once in the collect pass) routes a heavy reduction body
(fused_linear_jsd's ~7 live full-width fp32 tiles) from persistent [None] to looped [8192]
via a multi-tile spill ceiling (2.3-3.4x over the old persistent seed; n_spills 480->0).

welford — full-width M_BLOCK persist-cap (`_m_block_cap`) so a huge-M grid-size autotuner_min
raise cannot starve occupancy (only ever lowers an over-raised floor); fp32 per-chunk stat
accumulation fixes the pre-existing bf16/fp16 accuracy failures at wide reduction chunks.
welford config intentionally changes; the other 8 standard + 8 transfer kernels byte-identical.

+++ examples/welford.py
@@ -48,7 +48,10 @@ def welford(
         acc_m2 = torch.zeros_like(acc_cnt)
 
         for tile_n in hl.tile(n):
-            chunk = x[tile_m, tile_n]
+            # Accumulate per-chunk statistics in fp32 even for half-precision input: bf16/fp16 sum(chunk)
+            # and especially chunk*chunk drift with chunk width (NaN at fp16), so the fp32 online combine
+            # isn't enough alone. HBM load stays half (traffic unchanged); just an in-register promote.
+            chunk = x[tile_m, tile_n].to(torch.float32)
             # Count of VALID columns (the divisor): use the true valid count, not the constexpr
             # tile width (over-counts last-tile padding). Cast the mask to int32 BEFORE summing —
             # a bool operand gives CuTe a saturating accumulator (-> NaN).
